added animation handler to track animations on players and mobs
Used for dependOnCasterState in in region skills like echoing blades, and requireSkillCodes in additional effects like for Squall for players.
Used for pretty much every animation in mobs to get the timing correct. Required to properly schedule attack points in mob skill casts & to schedule mob skills that move the mobs.