Handling for `music_play_instrument_mastery` condition type
Handling of mastery updates has been changed to properly update the `music_play_instrument_mastery`.
This now enables the following Trophies to work properly:
- https://handbook.tadeucci.dev/trophies/23100142
- https://handbook.tadeucci.dev/trophies/23100143
- https://handbook.tadeucci.dev/trophies/23100144
- https://handbook.tadeucci.dev/trophies/23100168
- https://handbook.tadeucci.dev/trophies/23100169
- https://handbook.tadeucci.dev/trophies/23100170
- https://handbook.tadeucci.dev/trophies/23100178
- https://handbook.tadeucci.dev/trophies/23100179
- https://handbook.tadeucci.dev/trophies/23100180
- https://handbook.tadeucci.dev/trophies/23100197
- https://handbook.tadeucci.dev/trophies/23100198
- https://handbook.tadeucci.dev/trophies/23100199
- https://handbook.tadeucci.dev/trophies/23100224
- https://handbook.tadeucci.dev/trophies/23100225
- https://handbook.tadeucci.dev/trophies/23100226
- https://handbook.tadeucci.dev/trophies/23100232
- https://handbook.tadeucci.dev/trophies/23100233
- https://handbook.tadeucci.dev/trophies/23100234
- https://handbook.tadeucci.dev/trophies/23100235
- https://handbook.tadeucci.dev/trophies/23100236
- https://handbook.tadeucci.dev/trophies/23100237
- https://handbook.tadeucci.dev/trophies/23100254
- https://handbook.tadeucci.dev/trophies/23100255
- https://handbook.tadeucci.dev/trophies/23100256
- https://handbook.tadeucci.dev/trophies/23100277
- https://handbook.tadeucci.dev/trophies/23100278
- https://handbook.tadeucci.dev/trophies/23100310
- https://handbook.tadeucci.dev/trophies/23100311
- https://handbook.tadeucci.dev/trophies/23100353
- https://handbook.tadeucci.dev/trophies/23100354
- https://handbook.tadeucci.dev/trophies/23100355
- https://handbook.tadeucci.dev/trophies/23100410
- https://handbook.tadeucci.dev/trophies/23100411
- https://handbook.tadeucci.dev/trophies/23100412
- https://handbook.tadeucci.dev/trophies/23100413

Furthermore the `/player mastery exp` command has been updated to now add exp rather then set the
mastery value to the specified exp value as this causes issues for trophy tracking and is now more inline
with the actual description.

Additionally the `RewardGrade` for Trophies has been miscalculated if they had no reward but had more
additional grades that needed completing. Previously the `RewardGrade` had been increased past the
actual `Grade` of the Trophy to mark it as completed. This was only correct for trophies that did not
have additional `Grades` like e.g. music performance trophies causing the UI to skip to the next
`Grade` for display.

Refs: #484